+++ mobile/lib/common/routes.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get_10101/common/global_keys.dart';
 import 'package:get_10101/common/settings/channel_screen.dart';
+import 'package:get_10101/common/settings/delete_network_graph.dart';
 import 'package:get_10101/common/status_screen.dart';
 import 'package:get_10101/features/wallet/domain/destination.dart';
 import 'package:get_10101/features/wallet/send/send_lightning_screen.dart';
@@ -115,6 +116,14 @@ GoRouter createRoutes() {
                 parentNavigatorKey: rootNavigatorKey,
                 builder: (BuildContext context, GoRouterState state) {
                   return const StatusScreen();
+                },
+              ),
+              GoRoute(
+                path: DeleteNetworkGraphScreen.subRouteName,
+                // Use root navigator so the screen overlays the application shell
+                parentNavigatorKey: rootNavigatorKey,
+                builder: (BuildContext context, GoRouterState state) {
+                  return const DeleteNetworkGraphScreen();
                 },
               ),
               GoRoute(