+++ mobile/native/src/api.rs
@@ -21,6 +21,7 @@ use crate::trade::order::api::Order;
 use crate::trade::position;
 use crate::trade::position::api::Position;
 use crate::trade::users;
+use anyhow::anyhow;
 use anyhow::ensure;
 use anyhow::Context;
 use anyhow::Result;
@@ -30,6 +31,10 @@ use commons::OrderbookRequest;
 use flutter_rust_bridge::frb;
 use flutter_rust_bridge::StreamSink;
 use flutter_rust_bridge::SyncReturn;
+use lightning::util::persist::KVStore;
+use lightning::util::persist::NETWORK_GRAPH_PERSISTENCE_KEY;
+use lightning::util::persist::NETWORK_GRAPH_PERSISTENCE_PRIMARY_NAMESPACE;
+use lightning::util::persist::NETWORK_GRAPH_PERSISTENCE_SECONDARY_NAMESPACE;
 use ln_dlc_node::channel::UserChannelId;
 use rust_decimal::prelude::FromPrimitive;
 use rust_decimal::Decimal;
@@ -239,6 +244,18 @@ pub async fn get_positions() -> Result<Vec<Position>> {
         .collect::<Vec<Position>>();
 
     Ok(positions)
+}
+
+pub fn delete_network_graph() -> Result<()> {
+    crate::state::get_storage()
+        .ln_storage
+        .remove(
+            NETWORK_GRAPH_PERSISTENCE_PRIMARY_NAMESPACE,
+            NETWORK_GRAPH_PERSISTENCE_SECONDARY_NAMESPACE,
+            NETWORK_GRAPH_PERSISTENCE_KEY,
+            false,
+        )
+        .map_err(|e| anyhow!("{e:#}"))
 }
 
 pub fn subscribe(stream: StreamSink<event::api::Event>) {